+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 
+# Unreleased
 
+- Various documentation additions & improvements.
+- Add the [singleton example](./examples/singleton.rs), which shows a way to use `Block`s for something other than IPC.
 
 # 0.1.0